fixed a few warnings

+++ src/ParticleContainer.h
@@ -2,13 +2,13 @@
 #include "Particle.h"
 #include "common.h"
 #include <vector>
-#include "glm\glm.hpp"
+#include <glm\glm.hpp>
 #include "Grid.h"
 class ParticleContainer
 {
 public:
 	int MAXPARTICLES;
-	void updateParticles(float dt);
+	void updateParticles(int dt);
 	void findNeighbors(Particle &p, int pIndex);
 	ParticleContainer(int MAXPARTICLES);
 	void getPositions(std::vector<GLfloat> &positions);
